Make plugin dynamic (Component -> Listener)
